requirements.txt
Flask
simplejson
flask-sqlalchemy
pymysql
